requirements-dev.txt
# Development/test dependencies
pytest==9.0.2
pytest-django==4.11.1
coverage==7.13.0
pytest-cov==7.0.0
black==25.12.0
isort==7.0.0
flake8==7.3.0
mypy==1.19.0
django-debug-toolbar==6.1.0
django-stubs==5.2.8
djangorestframework-stubs==3.16.6
celery-stubs==0.1.3
